fix(nvml-mock): avoid unchecked int->uint32 conversion in clocks command
CodeQL flagged the uint32(mhz) narrowing of strconv.Atoi output in the
`clocks` command as an unchecked conversion, even though singleIntArg
bounds mhz to [0, 100000]. Change ClocksPatch to accept an int so there
is no narrowing at the call site; the generic overlay map carries the int
and MergeDeviceConfig's JSON round-trip decodes it into the uint32 schema
field.

+++ cmd/nvml-mock-ctl/main.go
@@ -222,7 +222,7 @@ func mutate(cmd, overlayPath, gpu, mode string, afterCalls int, xid uint64,
 			fprintf(stderr, "%v\n", perr)
 			return 2
 		}
-		if code := applyPatch(doc, target, base, mockctl.ClocksPatch(uint32(mhz)), stderr); code != 0 {
+		if code := applyPatch(doc, target, base, mockctl.ClocksPatch(mhz), stderr); code != 0 {
 			return code
 		}
 	case "throttle":
